Add model classes for Food, Entree, and Listable
This commit adds the model classes for Food, Entree, and Listable. The Food class represents a food item with its name, wholesale cost, retail cost, and an associated image. The Entree class represents an entree with a name and a list of Food objects. The Listable interface defines the common methods for getting the name, wholesale cost, retail cost, and checking equality for objects that can be listed.

These classes will be used to manage food items and entrees in the application.